Create README.md
Problem Description:
ABC Pharma is facing a significant challenge in understanding the persistency of drug usage as per physician prescriptions. The persistency of drug usage refers to the extent to which patients continue to take their prescribed medications over a specific period of time.
Currently, ABC Pharma Company relies on manual methods to track and analyze drug persistency. This involves reviewing patient records, conducting surveys, and relying on self-reporting, which can be time-consuming, prone to errors, and lack real-time insights.
To address this challenge, ABC Pharma Company has decided to approach an analytics company to automate the process of identifying drug persistency. The goal is to develop a data-driven solution that can accurately and efficiently track the usage of prescribed medications by patients, enabling ABC Pharma Company to gain valuable insights into the patterns of medication adherence.

Data Understanding:
The target variable in the dataset is the persistency flag. This flag indicates whether a patient was persistent or not in taking their prescribed medication. It serves as the target variable for the analysis, with values such as ‘persistent’ or ‘non-persistent’. Hence the persistency flag is the dependent variable which means other factors determine whether a patient was persistent or non-persistent with the drug.

Adherence refers to the extent to which patients follow their prescribed medication regimens. It is an important factor in accessing the effectiveness of treatments and
patient outcomes. The adherence-related features in the dataset provide insights into the patient’s medication adherence behaviors.

Understanding adherence patterns and factors affecting adherence is crucial for
pharmaceutical companies like ABC Pharma to evaluate the effectiveness of their drugs and develop strategies to improve patient adherence. By analyzing the
adherence-related features, ABC Pharma can gain insights into medication-taking behaviors, identify potential barriers to adherence, and tailor interventions to
improve patient compliance and persistency.